If/Else practice
Receive user input for two numbers. If the product of the numbers is less than or equal to 1000, return the product. If the product of the numbers is greater than 1000, return the sum.
Write a program to iterate through a range of 10 numbers and print the sum of the current and previous numbers.
Iterate through a word input from user and print all even numbered indexes.
Iterate through a list and print all items that are divisible by 5.